refactored to a new custom hook
