+++ src/components/ArticleHeadings/ArticleHeadings.js
@@ -1,7 +1,6 @@
-import React, { useState, useEffect } from 'react'
-import { useLocation } from '@reach/router'
+import React from 'react'
 import { sluggify } from '../Markdown/utils'
-import { scrollToTargetAdjusted } from '../../utils/utils'
+import { scrollToTargetAdjusted, usePageHash } from '../../utils/utils'
 import cx from 'classnames'
 import styles from './ArticleHeadings.module.scss'
 
@@ -35,15 +34,7 @@ const TOPICS = {
 }
 
 const ArticleHeadings = ({ list = [], crumbs = [] }) => {
-  const { hash } = useLocation()
-  const [pageHash, setPageHash] = useState()
-  useEffect(() => {
-    if (hash) setPageHash(hash)
-    const hashChangeHandler = ({ detail }) => setPageHash(detail)
-    window.addEventListener('hashScrollChanged', hashChangeHandler, false)
-    return () =>
-      window.removeEventListener('hashScrollChanged', hashChangeHandler, false)
-  }, [])
+  const pageHash = usePageHash()
   const topic = crumbs.length > 1 && crumbs[1].crumbLabel
   const appLink = topic && TOPICS[topic]
